+++ Commands/CmdSearch.cs
@@ -51,7 +51,7 @@ namespace MCForge.Commands
             {
                 string blocks = "";
                 bool mode = true;
-                for (byte i = 0; i < Block.maxblocks; i++)
+                for (ushort i = 0; i < Block.maxblocks; i++)
                 {
                     if (Block.Name(i).ToLower() != "unknown")
                     {
